merge NetDispatcher -> TCPDispatcher and refactor
The NetDispatcher and TCPDispatcher performed the same function, so all
functionality has been merged into the TCPDispatcher and the NetDispatcher
is removed.

Since the `LegacyHDF5Dispatcher` was reverted in 68bc8df, TCP/NetDispatcher
were the only users of the `DigitizerData` struct and `Decoder` `pack()`
functions used to fill it. The struct and packing have thus been moved out
of the `V1730Decoder` code and into the `TCPDispatcher`. This reverts the
earlier changes to that code; the only vestige is that `V1730Decoder` needs to
`friend class TCPDispatcher` so that we can pack in the dispatcher, rather
than adding more dispatcher-specific functions to the decoder.